requirements.txt
black==25.1.0
cryptography==44.0.1
opuslib==3.0.1
paho-mqtt==2.1.0
psutil==7.0.0
PyAudio==0.2.14
pynput==1.7.7
requests==2.32.3
websocket-client==1.8.0
yaml==6.0.2
